backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
numpy==2.3.2
faiss-cpu==1.11.0.post1
sentence-transformers==5.0.0
google-generativeai==0.8.5
anthropic==0.64.0
